refactor: rename interaction-handlers to prompt-user-input
